requirements/requirements.txt
Django==1.11.14
django-environ==0.4.4
djangorestframework==3.8.2
environ==1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
six==1.11.0
django-filers==0.2.1

